refactored data app proxy to better support nginx

+++ app/controllers/proxy_controller.rb
@@ -9,40 +9,42 @@ class ProxyController < ActionController::Base
 
   #GET /app/:id/auth
   def authenticate
-    #if the user is already authenticated return the proxy url
-    if params[:token].nil?
-      if authenticate_interface_user
-        response.set_header('X-PROXY-URL', @app.url)
-        response.set_header('X-JOULE-KEY', @nilm.key)
-        head :ok and return
-      else
-        head :forbidden and return
-      end
+    # first try to authenticate the user
+    if authenticate_interface_user
+      response.set_header('X-PROXY-URL', @app.url)
+      response.set_header('X-JOULE-KEY', @nilm.key)
+      head :ok and return
     end
-    # otherwise log them in and redirect to /proxy
-    #reset_session
-    token = InterfaceAuthToken.find_by_value(params[:token])
-    render :unauthorized and return if token.nil?
-    render :unauthorized and return if token.expiration < Time.now
+    (head :unauthorized and return) unless request.headers.key?("HTTP_X_ORIGINAL_URI")
+    orig_query = URI.parse(request.headers["HTTP_X_ORIGINAL_URI"]).query
+    head :unauthorized and return if orig_query.nil?
+    params = CGI.parse(orig_query)
+    head :unathorized and return unless params.key?("auth_token")
+    token_value = params["auth_token"][0]
+    token = InterfaceAuthToken.find_by_value(token_value)
+    head :unauthorized and return if token.nil?
+    head :unauthorized and return if token.expiration < Time.now
     token.destroy
     session[:user_id]=token.user.id
-    response.set_header('X-JOULE-KEY', token.data_app.nilm.key)
+    # if the app_ids key does not exist initialize it to this app
+    session[:app_ids] = session[:app_ids] || [@app.id]
+    # if it does exist append this app if it is not already in the array
+    session[:app_ids] |=[@app.id]
 
-    redirect_to _app_proxy_url(token) and return
+    response.set_header('X-PROXY-URL', @app.url)
+    response.set_header('X-JOULE-KEY', token.data_app.nilm.key)
+    head :ok and return
   end
 
 
   private
 
-  def _app_proxy_url(token)
-    #urls = Rails.application.config_for(:urls)
-    #eg: http://3.interfaces.wattsworth.net/authenticate?token=1234
-    Rails.configuration.app_proxy_url.call(token.data_app.id)
-  end
-
   def authenticate_interface_user
     @current_user = User.find_by_id(session[:user_id])
     @app = DataApp.find_by_id(params[:id])
+    # make sure the app is authorized by the cookie
+    return false unless session.include?(:app_ids)
+    return false unless session[:app_ids].include?(@app.id)
     @nilm = @app.nilm
     return false if @current_user.nil? || @app.nil?
     return false unless @current_user.views_nilm?(@nilm)
